Redacted OsuParser and Restructured workflow
I think OsuParser would be redundant, i'll leave most of the coding to the user instead of presenting an easily accessible *but* non-flexible workflow.

I'll need to relook at how the 2DVector to CSV would function and see if I can streamline it with a transpose since data is usually provided in columns instead of rows.